+++ NEWS.md
@@ -1,15 +1,23 @@
 
+## KernelKnn 1.0.4
+
+I added a try-catch Rcpp function to make possible the calculation of singular covariance matrices as sugggested in https://github.com/mlampros/KernelKnn/issues/1
+
+
 ## KernelKnn 1.0.3
 
-reimplementation of the Rcpp function due to ASAN-memory-errors
+Reimplementation of the Rcpp function due to ASAN-memory-errors
+
 
 ## KernelKnn 1.0.2
 
 I updated the Description file with a URL and a BugReports web-address.
 
+
 ## KernelKnn 1.0.1
 
 Currently, Software platforms like OSX do not support openMP, thus I've made openMP optional for all cpp functions.
+
 
 ## KernelKnn 1.0.0